Adding Treasury Features to the App

+++ StoreApplication/StoreApplication/Models/TreasuryAccount.cs
@@ -1,10 +1,11 @@
 ﻿namespace StoreApplication.Models
 {
-    public class TreasuryAccount
+    public class TreasuryAccount 
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Balance { get; set; }
+        public double Balance { get; set; }
+        
         public List<TreasuryTransaction> Transactions { get; set; }
     }
 }
